python-bot/requirements.txt
python-dotenv==1.0.1
pymysql==1.1.1
SQLAlchemy==2.0.38
requests==2.32.3
websocket-client==1.8.0